Example Dataset: data_fukuchi_2018

An example dataset of joint angles from Fukuchi et al., 2018, released under https://creativecommons.org/licenses/by/4.0/

Usage

data_fukuchi_2018

Format

A data frame with 1375620 rows and 6 variables:

stride

Numeric stride

participant

Character participant

task

Character task

variable

Character variable

value

Numeric value

gait_type

Character gait_type

Source

https://figshare.com/articles/dataset/A_public_data_set_of_overground_and_treadmill_walking_kinematics_and_kinetics_of_healthy_individuals/5722711

References

Fukuchi CA, Fukuchi RK, Duarte M. 2018. A public dataset of overground and treadmill walking kinematics and kinetics in healthy individuals. PeerJ 6:e4640 https://doi.org/10.7717/peerj.4640