Visualize OpenSim Model

This function visualizes an OpenSim model using the `opensim-cmd` tool.

Usage

osim_viz_model_cmd(
  opensim_path = "C:/\"OpenSim 4.5\"/bin",
  model_file_path = "Trunk_pelvis_LL.osim"
)

Arguments

opensim_path

A character string specifying the path to the OpenSim bin directory.

model_file_path

A character string specifying the path to the model file (.osim).

Value

The result of running the OpenSim command.